Created ability to simulate phenotypes

+++ functions/simulation_helpers/simulate_GRM_phenos.py
@@ -8,6 +8,7 @@ Created on Fri Sep 23 10:00:54 2022
 
 @author: christian
 """
+import os
 import numpy as np
 from scipy.stats import random_correlation
 
@@ -15,24 +16,26 @@ from scipy.stats import random_correlation
 #%% Simulate/load a random GRM (nonsparse)
 
 def sim_GRM(n, GRM_save) :
-    # Generate a random rorrelation matrix (it needs sum of eigenvalues to equal dimensionality)
-    eigs = np.random.rand(n)
-    eigs = eigs / sum(eigs) * n
-    A = random_correlation.rvs(eigs, tol = 1e-1)
-    # Scale correlation into covariance 
-    np.save(GRM_save + ".npy", A)
+    if not os.path.exists(GRM_save) :
+    
+        # Generate a random rorrelation matrix (it needs sum of eigenvalues to equal dimensionality)
+        eigs = np.random.rand(n)
+        eigs = eigs / sum(eigs) * n
+        A = random_correlation.rvs(eigs, tol = 1e-1)
+        # Scale correlation into covariance 
+        np.save(GRM_save + ".npy", A)
 
 
 def simulate_phenotypes(GRM, df, sigmas, sim_prefix) :
-    num_sites = len(np.unique(df.Zs))
+    num_sites = len(np.unique(df.abcd_site))
     n = GRM.shape[0]
 
     
     for i, s in enumerate(sigmas) :
         # Generate random site effects
         Bs = np.random.normal(0,  s[1], num_sites)
-        BBs =  {i : Bs[i] for i in range(num_sites)}
-        site_effects = [BBs[site] for site in df.Zs]
+        BBs =  {"site" + str(i + 1).rjust(2, "0") : Bs[i] for i in range(num_sites)}
+        site_effects = [BBs[site] for site in df.abcd_site]
         
         # compute covariance for simulate(V) for simulation
         V = s[0] * GRM + s[2] * np.identity(n)
